+++ packages/cinterop/src/commonMain/kotlin/io/realm/kotlin/internal/interop/sync/ProgressDirection.kt
@@ -1,0 +1,6 @@
+package io.realm.kotlin.internal.interop.sync
+
+expect enum class ProgressDirection {
+    RLM_SYNC_PROGRESS_DIRECTION_UPLOAD,
+    RLM_SYNC_PROGRESS_DIRECTION_DOWNLOAD,
+}