+++ packages/cinterop/src/jvm/jni/java_class_global_def.hpp
@@ -61,6 +61,7 @@ private:
         , m_io_realm_kotlin_internal_interop_sync_after_client_reset_handler(env, "io/realm/kotlin/internal/interop/SyncAfterClientResetHandler", false)
         , m_io_realm_kotlin_internal_interop_core_error_utils(env, "io/realm/kotlin/internal/interop/CoreErrorUtils", false)
         , m_io_realm_kotlin_internal_interop_sync_async_open_callback(env, "io/realm/kotlin/internal/interop/AsyncOpenCallback", false)
+        , m_io_realm_kotlin_internal_interop_progress_callback(env, "io/realm/kotlin/internal/interop/ProgressCallback", false)
     {
     }
 
@@ -82,6 +83,7 @@ private:
     jni_util::JavaClass m_io_realm_kotlin_internal_interop_sync_after_client_reset_handler;
     jni_util::JavaClass m_io_realm_kotlin_internal_interop_core_error_utils;
     jni_util::JavaClass m_io_realm_kotlin_internal_interop_sync_async_open_callback;
+    jni_util::JavaClass m_io_realm_kotlin_internal_interop_progress_callback;
 
     inline static std::unique_ptr<JavaClassGlobalDef>& instance()
     {
@@ -182,6 +184,11 @@ public:
     inline static const jni_util::JavaClass& async_open_callback()
     {
         return instance()->m_io_realm_kotlin_internal_interop_sync_async_open_callback;
+    }
+
+    inline static const jni_util::JavaClass& progress_callback()
+    {
+        return instance()->m_io_realm_kotlin_internal_interop_progress_callback;
     }
 
     inline static const jni_util::JavaMethod function0Method(JNIEnv* env) {