+++ packages/jni-swig-stub/src/main/jni/realm_api_helpers.h
@@ -98,6 +98,9 @@ sync_before_client_reset_handler(realm_sync_config_t* config, jobject before_han
 void
 sync_after_client_reset_handler(realm_sync_config_t* config, jobject after_handler);
 
+void
+realm_sync_session_progress_notifier_callback(void *userdata, uint64_t transferred_bytes, uint64_t total_bytes);
+
 // Explicit clean up method for releasing heap allocated data of a realm_value_t instance
 void
 realm_value_t_cleanup(realm_value_t* value);