+++ packages/library-sync/src/commonMain/kotlin/io/realm/kotlin/mongodb/internal/SyncSessionImpl.kt
@@ -17,6 +17,7 @@
 package io.realm.kotlin.mongodb.internal
 
 import io.realm.kotlin.internal.RealmImpl
+import io.realm.kotlin.internal.interop.sync.ProgressDirection
 import io.realm.kotlin.internal.interop.RealmInterop
 import io.realm.kotlin.internal.interop.RealmSyncSessionPointer
 import io.realm.kotlin.internal.interop.SyncSessionTransferCompletionCallback
@@ -26,9 +27,18 @@ import io.realm.kotlin.internal.interop.sync.SyncErrorCode
 import io.realm.kotlin.internal.interop.sync.SyncErrorCodeCategory
 import io.realm.kotlin.internal.platform.freeze
 import io.realm.kotlin.internal.util.Validation
+import io.realm.kotlin.mongodb.sync.Direction
+import io.realm.kotlin.mongodb.sync.Progress
+import io.realm.kotlin.mongodb.sync.ProgressMode
 import io.realm.kotlin.mongodb.sync.SyncSession
 import kotlinx.coroutines.TimeoutCancellationException
+import kotlinx.coroutines.channels.BufferOverflow
 import kotlinx.coroutines.channels.Channel
+import kotlinx.coroutines.channels.awaitClose
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.buffer
+import kotlinx.coroutines.flow.callbackFlow
+import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.withContext
 import kotlinx.coroutines.withTimeout
 import kotlin.time.Duration
@@ -71,6 +81,43 @@ internal open class SyncSessionImpl(
 
     override fun resume() {
         RealmInterop.realm_sync_session_resume(nativePointer)
+    }
+
+    @Suppress("invisible_member", "invisible_reference")
+    override fun progress(
+        direction: Direction,
+        progressMode: ProgressMode,
+    ): Flow<Progress> {
+        val callbackFlow: Flow<Progress> = callbackFlow {
+            val token = RealmInterop.realm_sync_session_register_progress_notifier(
+                nativePointer,
+                when (direction) {
+                    Direction.DOWNLOAD -> ProgressDirection.RLM_SYNC_PROGRESS_DIRECTION_DOWNLOAD
+                    Direction.UPLOAD -> ProgressDirection.RLM_SYNC_PROGRESS_DIRECTION_UPLOAD
+                },
+                progressMode == ProgressMode.INDEFINITELY
+            ) { transferredBytes: Long, totalBytes: Long ->
+                val progress = Progress(transferredBytes.toULong(), totalBytes.toULong())
+                trySend(progress)
+                if (progressMode== ProgressMode.CURRENT_CHANGES && progress.isTransferComplete) {
+                    close()
+                }
+            }
+            awaitClose {
+                RealmInterop.realm_sync_session_unregister_progress_notifier(nativePointer, token)
+            }
+        }
+        return callbackFlow
+            .onEach {
+                // If we are done downloading data ensure that the realm is up to date as users
+                // would expect to see all the downloaded data
+                if (it.isTransferComplete) {
+                    if (direction == Direction.DOWNLOAD) {
+                        realm?.refresh()
+                    }
+                }
+            }
+            .buffer(1, BufferOverflow.DROP_OLDEST)
     }
 
     /**