+++ packages/library-sync/src/commonMain/kotlin/io/realm/kotlin/mongodb/sync/Direction.kt
@@ -1,0 +1,18 @@
+package io.realm.kotlin.mongodb.sync
+
+/**
+ * A **direction** indicates whether a given [Progress]-flow created with
+ * [SyncSession.progress] is reporting changes in ongoing uploads or downloads.
+ */
+public enum class Direction {
+    /**
+     * Used to pass to [SyncSession.progress] to create a flow that reports [Progress]
+     * in the [SyncSession]'s downloads.
+     */
+    DOWNLOAD,
+    /**
+     * Used to pass to [SyncSession.progress] to create a flow that reports [Progress]
+     * in the [SyncSession]'s uploads.
+     */
+    UPLOAD,
+}