+++ packages/library-sync/src/commonMain/kotlin/io/realm/kotlin/mongodb/sync/SyncSession.kt
@@ -18,6 +18,10 @@ package io.realm.kotlin.mongodb.sync
 
 import io.realm.kotlin.Realm
 import io.realm.kotlin.mongodb.exceptions.SyncException
+import io.realm.kotlin.mongodb.sync.SyncSession.ErrorHandler
+import io.realm.kotlin.mongodb.sync.SyncSession.State.ACTIVE
+import io.realm.kotlin.mongodb.sync.SyncSession.State.DYING
+import kotlinx.coroutines.flow.Flow
 import kotlin.time.Duration
 
 /**
@@ -39,6 +43,11 @@ import kotlin.time.Duration
  * The [SyncSession] object is thread safe.
  */
 public interface SyncSession {
+
+    /**
+     * The current session state. See [State] for more details about each state.
+     */
+    public val state: State
 
     /**
      * Calling this method will block until all known remote changes have been downloaded and
@@ -69,11 +78,6 @@ public interface SyncSession {
     public suspend fun uploadAllLocalChanges(timeout: Duration = Duration.INFINITE): Boolean
 
     /**
-     * The current session state. See [State] for more details about each state.
-     */
-    public val state: State
-
-    /**
      * Pauses synchronization with Atlas until the Realm is closed and re-opened again.
      *
      * Synchronization can also be re-activated by calling [resume].
@@ -93,6 +97,22 @@ public interface SyncSession {
      * If the session state is [State.DYING], the session will be moved back to [State.ACTIVE].
      */
     public fun resume()
+
+    /**
+     * Create a [Flow] of [Progress]-events that tracks transfers of the [SyncSession].
+     *
+     * If the flow is created with [ProgressMode.CURRENT_CHANGES] the [Progress] will
+     * only ever increase and will complete once `Progress.isTransferComplete = true`.
+     *
+     * If the flow is created with [ProgressMode.INDEFINITELY] the [Progress] can both
+     * increase and decrease since more changes might be added while the flow is still active. This
+     * means that it is possible for one [Progress] instance to report
+     * `isTransferComplete = true` and subsequent instances to report `isTransferComplete = false`.
+     */
+    public fun progress(
+        direction: Direction,
+        progressMode: ProgressMode,
+    ): Flow<Progress>
 
     /**
      * Interface used to report any session errors.